requirements.txt
flake8==6.1.0
looker-sdk==23.16.0
python-dotenv==1.0.0